pulled clip and unet into seperate files
